Move contents of `crate::arena` into their own sub-modules
